wip: added keypad, don't know what to do with display yet, reading bytes from the rom just fine now

+++ src/consts.rs
@@ -1,8 +1,8 @@
 pub const DISPLAY_WIDTH: usize = 64;
 pub const DISPLAY_HEIGHT: usize = 32;
-pub const ROM_LOCATION: usize = 0x200; // because the system expects the app to be loaded at 0x200
-pub const FONT_LOCATION: usize = 0x80; // fonts are expected at 0x80
-pub const FONT_SET: [usize; 80] = [
+pub const ROM_LOCATION: u16 = 0x200; // because the system expects the app to be loaded at 0x200
+pub const FONT_LOCATION: u8 = 0x80; // fonts are expected at 0x80
+pub const FONT_SET: [u8; 80] = [
     0xF0, 0x90, 0x90, 0x90, 0xF0, // 0
     0x20, 0x60, 0x20, 0x20, 0x70, // 1
     0xF0, 0x10, 0xF0, 0x80, 0xF0, // 2
